fix(trusty-mpm): split resolver for line-cap + align project_resolve response shape + stronger tests (refs #1517)
- Split monolithic resolver.rs (641 SLOC > 500 cap) into resolver/mod.rs
  (production code, ~340 SLOC) and resolver/tests.rs (26 tests, ~330 SLOC).
  Uses `#[path = "tests.rs"] mod tests;` so Cargo finds the sibling file.
  `bash scripts/check_line_cap.sh` now reports 0 violations.

- Fix doc/contract mismatch in mcp_project.rs: the project_resolve doc-comment
  showed confidence/reason at the top level of the JSON response, but the
  actual json!() macro nests them inside the primary object. Updated the
  doc-comment to show the real nested shape.

- Strengthen dispatch_project_resolve_no_match test: replaced the weak
  `text.contains("null") || text.contains("error")` assertion with a proper
  JSON parse + structural assertions: primary must be Value::Null and the
  error field must contain the original query string.

- Add doc note to keyword_score() documenting the intentional description-only
  dead-zone (score 0.2 < KEYWORD_COLLECTION_FLOOR 0.3, so desc-only hits
  never surface alone; Phase 2 scope if needed).
